requirements.txt
yfinance==0.2.44
pandas==2.2.2
numpy==2.0.1
Jinja2==3.1.4
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
httpx==0.27.2
python-dotenv==1.0.1
flask>=3.0.0
openai
tqdm
requests
feedparser>=6.0.0
pyinstaller>=6.0.0